requirements.txt
Flask==1.1.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
python-socketio==4.3.1
